+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: nloomans <nloomans@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/03/25 11:51:37 by nloomans       #+#    #+#                */
-/*   Updated: 2019/03/28 20:11:39 by nloomans      ########   odam.nl         */
+/*   Updated: 2019/03/28 20:17:41 by nloomans      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned long	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
