+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: nloomans <nloomans@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/03/21 15:35:12 by nloomans       #+#    #+#                */
-/*   Updated: 2019/03/28 20:12:15 by nloomans      ########   odam.nl         */
+/*   Updated: 2019/03/28 20:18:48 by nloomans      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strncat(char *s1, const char *s2, size_t n)
 {
-	unsigned long	i;
-	size_t			s1_len;
+	size_t	i;
+	size_t	s1_len;
 
 	i = 0;
 	s1_len = ft_strlen(s1);
